requirements.txt
discord.py
pymongo
dnspython
praw